requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
openai>=1.3.0
websockets>=12.0

# For Python 3.13, we need to install these separately with pre-built wheels
# Run these commands manually if the script fails:
# pip install --only-binary :all: numpy openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
duckduckgo-search>=5.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
scikit-learn>=1.3.0
